Separating palette from game.
